♻️ Outsource types
